Fix:Capture image with different resolution and format.
:Release Note:
Capturing image with a resolution and format that is different
from preview resolution and format.

:Detailed Notes:
If preview parameters and capture parameters are different then
set the capture parameters to capture image and reset it to
preview parameters after capture is completed.

:Testing Performed:
Tested on raspbery pi3 boards
QR-255 - Verify image capture with different parameters

:QA Notes:
NA

:Issues Addressed:
[PLAT-64422] Unable to capture if resolution differs from preview format

Change-Id: I91a5ddf028b19c80c40b1c3c2dd73a491f8bec03