Fix Camera service crash
:Release Notes:
Fix camera service crash with AN VC-500 camera

:Detailed Notes:
Handled the frame size for all types of camera.
Made the frame size to copy generic based on width and height.
The crash was happening as the frame size returned by device was more
than the allocated memory.
The frame size was calculated with 2 bytes per pixel which actually
should be 4.

:Testing Performed:
Built and Tested on RPI.
Verified test cases:QO-3777,QO-3784,QO-3829,QR-741

:QA Notes:
NA

:Issues Addressed:
[PLAT-74619] Camera service crash with AN VC-500 camera

Change-Id: Iddcc79e149ffdfcfb430d9067a647569b594b7cd
Reviewed-on: http://gpro.lgsvl.com/231845
Reviewed-by: Gururaj Patil <gururaj.patil@lge.com>
Tested-by: Gururaj Patil <gururaj.patil@lge.com>
Reviewed-by: Rajeev Agrawal <rajeev.agrawal@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>