Add BUILD_UTILS_ONLY option in CMakeLists
:Release Notes:
Added a new option to build only camera utilities.

:Detailed Notes:
A new CMake option called BUILD_UTILS_ONLY was added.
When it is set to ON, only the src/common directory will be built.
If it is not set, the build process will include more directories:
src/plugins, src/services, and tests,and it will run the conditional
logic to build system bus files based on the USE_NEW_ACG variable.

:Testing Performed:
Local Build & Test : OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-13814] Apply clang build for cambufferlib and camera-utils

Change-Id: Ie82ea9429b81dde88e590b0ca4ec8677e3309a6a