camera-utils: Add CMake configuration and Apache License
:Release Notes:
Introduce CMake configuration for building and add Apache License.

:Detailed Notes:
This commit updates the CMakeLists.txt in the common directory
to specify a minimum required version and set the C++ standard to 17.
It also includes necessary modules and adds include directories for
GLib2 and camera headers.
Additionally, it adds the Apache License 2.0 to the
LICENSE file and introduces new header files for shared memory and Luna
service client functionalities under the camera directory.

:Testing Performed:
Local Build & Test : OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-16693] [OSE-RPI4]: Unable to create .sh file

Change-Id: I84e569f41564e1b35b3acd67cf3a8fef2db07d03